% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importcsv.R
\name{BasicPack}
\alias{BasicPack}
\title{BasicPack Function}
\usage{
BasicPack()
}
\description{
BasicPack is the first step towards making data science more userfriendly and easie.This BasicPack automatecally detact and import train and test dataset(only of csv format)
}
\examples{
BasicPack()
}
\keyword{importcsv}

